requirements.txt
pdfplumber==0.10.3
nltk
scikit-learn
langchain==0.1.20
langchain-community==0.0.38
openai==1.30.1
mistralai==0.1.8
tiktoken==0.6.0
streamlit
python-dotenv==1.0.1
PyYAML==6.0.1
tenacity==8.2.3
requests==2.32.3
pandas==2.2.2